refactor: extract substream modules
